Add macOS support to build-wheels script
Modify build-wheels script to also support for macOS. Although this is
essentially (for now) a glorified wrapper around build-wheel-mac.sh, it
provides a single entry point for building wheels, which should simplify
things for users, and also for CI.

This will also provide a separate entry point for tests, which has not
yet been implemented. Wheel extraction has also not been implemented,
although that is less important on macOS since the wheel is built
directly on the host.