Refactor wheel build script(s)
Refactor build-wheels script, moving the logic into a "package". This
allows us to do clever things like have the Linux and macOS bits in
separate files, and treat those as objects with a shared interface, so
that we only need to check at one place what platform we are on.

The front-end is a little ugly, but works.