Fix determination of macOS wheel platform tag
Add custom logic to try to guess the macOS wheel platform tag. We really
ought to be getting this from setuptools or such, but unfortunately the
macOS logic is buried in the unexposed bowels of bdist_wheel and is not
even directly usable. (As a bonus, however, this eliminates a dependency
on deprecated distutils.)