Remove determination of macOS wheel platform tag
As best I can tell, there is no accessible, reliable way to determine
the wheel platform tag on macOS (the method specified by PEP 425 does
not give the same answer as the logic buried in the unexposed bowels of
wheel.bdist_wheel that is used to generate the actual wheel name).
Since the "real" method is not only buried, but not even directly
usable, and there does not, after some effort, appear to be a good,
reliable alternative... give up and rely on the fact that we are
expecting only one wheel to be built, and just use glob to get its name.

(As a bonus, this eliminates a dependency on deprecated distutils.)