[examples/rimless_wheel] Improve meshcat animation
Previously, the floating-base pose that was sent to the visualizer
changed discontinuously at each step.  This is fine for forward
simulation, but in meshcat's animation the interpolation effects of
scrubbing through frames caused visual artifacts.  This PR simply
unwraps the pitch angle to make it continuous during a simulation.

It also adds some missing const specifier to the accessor functions
(which were needed to write the test).