[multibody] Filter collisions within welded subgraphs
Closes: #16422

This is a generalization of the the already-existing code to filter
collisions within bodies welded to the world.  Now, we filter collisions
within any welded subgraph of the model.

On the way to constructing a test, remove some unused infrastructure and
(mostly) prefer method calls to multiple optional constructor
parameters.

Also, back-fill a test that should have accompanied PR #11116.