Use clang 12
Remove installation of clang 9. Change all documentation and invocations
to use clang-12 (likewise with llvm and clang-format). Remove incorrect
documentation for running clang-format on macOS (the specified path is
incorrect, and in practice, the binary should be in $PATH). Remove
support for clang-6 and earlier from tools/lint/clang_format.py, which
greatly simplifies the logic.