[common] Improve symbolic polynomial performance via is_expanded
In cases where we are operating on affine expressions over variables,
notice cases where we know that is_expanded == true upon construction
so that we don't need to re-expand it later.

This yields a ~2x speedup in BenchmarkSosProgram1. As of this change,
all calls from benchmark program to Expression::Expand are no-ops;
the expression is always already known to be expanded.