"Fix" install RPATH for MOSEK
Modify how we adjust RPATHs in installed libraries to rewrite two types
of paths:

- Paths starting with `@loader_path`.
- Paths into Bazel's private build ares (`/private/var/tmp/_bazel_$USER`).

Previously, we only rewrote the former. However, for reasons which are
unclear, MOSEK gets linked using the latter, which was resulting in the
installed library containing a reference to the absolute path of MOSEK
as found in the build tree (which is clearly not useful).