Fix install RPATH mangling
Modify how we adjust RPATHs in installed libraries to rewrite paths to
libraries that are being installed, regardless if they are already
relative to `@loader_path`.

The reason for skipping already-absolute paths is unclear, since the
absolute location is unlikely to be the desired path for the installed
library. In particular, MOSEK, for reasons which are unclear, gets
linked using an absolute path into the bowels of Bazel's private build
area, which is certainly not going to exist on the machine of a user
installing from a distribution tarball.

This replaces a previous, somewhat sketchy change that was rushed into
production.