Refactor wheel build script(s)
Refactor build-wheels script, moving the logic into a "package". This
will ultimately simplify things if and when we want to support different
(i.e. non-Docker) paradigms for building wheels, by moving key
functionality into an abstracted layer that can be replaced by different
logic.

The front-end is a little ugly, but works.