[parsing] Warn about unsupported URDF features (#16953)
* [parsing] Warn about unsupported URDF features

Closes: #16783
Closes: #16181

Based on study of the ROS documentation and urdfdom module sources, add
warnings and documentation for ignored URDF features. Implementation is similar
to the in-progress mujoco parser, with an eye to future integration of
diagnostics there.

Adjust tests for existing files that emit new warnings, and add explicit
tests for each of the added warning types.

For Release Notes:
This patch modifies some installed example files to purge unsuppported
elements that now provoke warnings.