Add snippets to set the compiler, for CI
Create a pair of small CMake files that can contain the necessary
actions to set up the appropriate compiler for CI. This will allow us to
move that logic out of the CI repository, which will in turn allow us to
make changes (e.g. upgrading to Clang 12) without having to synchronize
them across two different repositories.

As of writing, this isn't used yet; it needs to land here first before
CI can start to make use of it.