Add Bazel config for compiler selection (#17043)
* Add Bazel config for compiler selection

Add ability to select which compiler to use by passing an appropriate
--config to Bazel. Update documentation accordingly.

In addition to being more convenient for users, this moves the
specification of what clang version to use (only relevant on Ubuntu) to
the Bazel configuration, rather than the user needing to specify this.
This will also allow this information to be removed from the CI
infrastructure.