Add tool to capture and report --config compiler
Create a tool to capture what compiler is being requested via
`--config={gcc,clang}`. This information is needed by the CI
infrastructure. (Specifically, CI needs a way to extract the potentially
version-dependent name of the Clang compiler that is selected by
`--config=clang`, so that this information does not need to also live
in the CI repository.)