Add a module for parsing `otool` output
Create a new Python module for parsing the output of (macOS) `otool`
into structured data. Use this in `installer.py`.

The new module supports extraction of load commands (`otool -l`, only
the "Load Command N" blocks) and linked libraries (`otool -L`), and
should be slightly more robust than the prior parsing code, particularly
the old load command parsing which acted on any "path" key without
regard to the type of command to which the key belonged. This will also
allow this logic to be reused for other tools that need to perform RPATH
manipulations.