[examples/rimless_wheel] Resolves subtle numerics in witness function logic (#17078)
* [examples/rimless_wheel] Resolves subtle numerics in witness function logic.

Resolves #8589. If the guard was triggered with exactly zero
penetration, then we could fail on the following step because the
wheel was positioned exactly in contact with the ground (and we are
using WitnessFunctionDirection::kPositiveThenNonPositive). The
unfortunate solution is to add machine epsilon to make the next
witness value strictly positive.

This does *not* implement the change to the double support threshold
suggested in #8589, where we use the accuracy from the context as the
threshold for detecting double support. I tried it, but it resulted in
a far too subtle interplay between the numerics of the witness
triggers and the detection of double support (and in practice led to
infinite loops). I've added an extra line of documentation to capture
that lesson, but do not consider it a priority to do more than that.