[math] Move DifferentiableNorm to math; use it in Wing.
As discussed in https://stackoverflow.com/a/72160960/7829525 the
`Wing` class implementation was using a vector `norm()` call that is
not differentiable at zero velocity.  For quadrotors, that's no good.

I found internal::DifferentiableNorm hidden in the
QuaternionIntegrationConstraint. This is a gem.  I've moved it to
`math`, and called it from `Wing`.