Use clang 12 (#16937)
Remove installation of clang 9. Change all documentation and invocations
to use clang-12 (likewise with llvm and clang-format). Remove incorrect
documentation for running clang-format on macOS (the specified path is
incorrect, and in practice, the binary should be in $PATH). Remove
support for clang-6 and earlier from tools/lint/clang_format.py, which
greatly simplifies the logic.

Also, fix some places where the minimum required GCC version did not get
updated, and fix some documentation that had outdated or missing
cross-references, which should help with keeping all the bits up to date
in the future.

Co-authored-by: Jeremy Nimmer <jeremy.nimmer@tri.global>