[examples] Provide default values for actuation input ports (#17145)
Previously, the AcrobotPlant example crashed if the input port was not
properly connected.  I've provided default a input value for it here.
And I've added them to PendulumPlant and CompassGait, too, for
consistency.  The compass gait example, in particular, really should
have had it... it's famous for being a passive dynamic walker!

Related to #17025.

Note: In jupyter (both deepnote and locally), the AcrobotPlant issue
was manifesting as a completely silent kernel crash, with absolutely
no hint/diagnositic.