[workspace] Tweak PETSc prefetch and xmm config
This is a step towards compiling on Apple M1.

Only use <xmmintrin.h> when __SSE__ is defined -- it is always defined
by default in our Ubuntu builds, but is not defined on M1 builds.

The __builtin_prefetch is valid to use on all of our supported platforms.
It allows the compiler to choose the best prefetch instruction.