[geometry/optimization] Make `IrisOptions.enable_ibex = false` by default.
Previously the default was to have Ibex enabled, but this has fallen
out of favor -- we find that Ibex cannot solve many problems of
interest.

In addition, we allow the test that requires Ibex to pass quietly if
IbexSolver is not available + enabled.  This is required for tests to
pass on mac arm64.