[visualization] DrakeVisualizer broadcasts new deformable messages
This is a step towards visualizing deformable geometries.

The topology of the deformable mesh is (re)analyzed when the geometry
version changes. A new message type that encodes the vertex positions
of the surface mesh of the deformable geometry is sent via a new
channel every time the "draw" messages are emitted for existing
geometries.

A couple of helper functions to identify deformable geometries are also
added in SceneGraphInspector.