Simplify compiler checks
Rewrite the logic in `cc/repository.bzl` to use a helper function for
testing if the compiler is of a supported version, based on a dictionary
of supported compilers. This replaces the previous logic which was
heavily redundant and required the supported compiler version to be
specified in multiple places.