Add macos_wheel platform
Add ability to override the platform on macOS. Add a new platform,
"macos_wheel", which roughly corresponds to "manylinux" on Ubuntu.
(There are some additional implications to "manylinux" specifically with
regard to additional processing of the resulting wheels, but
conceptually, both represent their respective platforms being used to
produce a wheel.)

This also removes `distribution = "manylinux"`; on careful inspection,
it turns out that logic would never have been executed. (We may end up
revisiting this, but for now, removing it is consistent with not having
a macOS equivalent, which would require additional changes. The same may
well be true for `distribution = "manylinux"`.)