Implement testing for macOS wheel
Modify macOS wheel build script to use a subdirectory under the build
root rather than /opt/python for the virtualenv used to build the wheel,
in order to avoid conflicts with the test environment. Add a new wrapper
script to set up the test environment on macOS and run the "real" test
script (i.e. the stuff we would do with Docker on Linux). Hook this up
in build-wheels. Hook up "extracting" the wheel on macOS (which is just
copying it from its spot in the build root into the user-specified
output directory).

This also changes the location of the Python virtual environment used
for testing in order to better avoid potential conflicts, which (unlike
on Linux, where everything runs in Docker) are a potential issue on
macOS.