[workspace] For sdformat and related, use gazebosim not ignition (#17240)
The ignitionrobotics organization is now empty; development has moved to
the gazebosim organization, so we should to update github_archive to point
to the new organization.  For the math and utils libraries, the repository
name also changed from ign to gz; that requires a checksum update as well,
because the source tarballs embed that name.

Upstream also no longer uses GitHub release announcements for minor releases.
Therefore, we need to use tags when checking for new releases.