[parsing] Restrict collision filters to models within a parse operation (#17241)
* [parsing] Restrict collision filters to models within a parse operation

Relevant to: #14785

This patch implements a stronger enforcement of model encapsulation for
collision filters. Previously, filters defined outside any model could refer to
the world body or parts of the default model by scoped names, or to default
model bodies by unqualified names. Now, those references are parse errors, and
using a model index from outside the parse as a parameter is a programming
error.

There might ever be a reason (I couldn't think of one) to define collision
filters against the world body; I am fine for that to involve a special
privileged name and be future work.

I couldn't think of any reason why a model file should be able to refer to
bodies in the default model, which (in the current implementation of Parser),
could never be populated from model files.