[viz] prototype of multi-role visualization
This patch is a prototype of achieving multi-role visualization via the
drake_visualizer message protocol. It uses additional robot_num allocations and
naming tricks to allow viewer-side control of what geometries are visible.

Fair warning: this patch "works" at a demo level, but is riddled with
questionable choices, and lacks tests.