[workspace] Build NLopt from source
Compared to the Ubuntu and macOS packages, this removes the LGPL'd
"luksan" solver from the build, which means we can link the library
statically (and eventually, hidden). This might mean that NLopt is
no longer able to solve certain kinds of programs, but that is a
fair trade-off given all of the problems caused by LGPL linking.

The package.BUILD.bazel is modeled on nlopt.BUILD.bazel in a10ebb7.

Deprecate the now-unused current (host OS) nlopt external.