[workspace] Clarify os_result.target (formerly distribution)
The os_result.distribution was documented to be fine-grained, with wheel
builds using a different value, but in 65d0bde we mistakenly reverted it
to "ubuntu" even for wheel builds.

Here we clarify the documentation, rename it to "target" for even more
clarity, and update using-code to match.

We also start to make room for the is_macos_wheel option, even though
it is neither set nor used yet.