Implement macos_wheel platform (#17301)
* Implement macos_wheel platform

Add support for the recently added "macos_wheel" target/platform.
Add ability to override the platform on macOS, which finishes
implementing the "macos_wheel" target/platform that was recently
introduced. Implement support for the same. This should, with a proper
environment, allow Drake to be built in a mode that can be used to
produce a wheel.

(Scripts to actually perform such a build, and particularly to establish
the aforementioned "proper environment", will be added in subsequent
commits.)

Also, explicitly add the pkg-config paths to Bazel's search paths to
allow us to find dependencies built for wheels. This is required as we
cannot inject these into /usr/local on macOS, as we would in the Ubuntu
Docker containers.