[workspace] Improve qhull hidden visibility
When building C++ code, we should not use -fvisibility=hidden because it
infects all #included code, even the standard library. This causes a lot of
console spam during macOS builds, and is a typeinfo (vtable) problem even on
Ubuntu, potentially leading to ODR violation. Instead, we should wrap the
third-party C++ code in a inline hidden namespace.  That way, only the
third-party code itself is hidden, not anything that it #includes.

We are substantially patching and subsetting this library to weave it into
Drake. We shouldn't give the illusion to users that it's available for reuse
downstream; therefore, we'll add an "_internal" suffix and deprecate the
original repository name.