[math] Fix SparseMatrixToTriplets to only permit sparse matrices
When it was given a dense matrix by mistake, it included all of the
zeros in the returned triplets.

Fix OsqpSolver to not try to pass a dense matrix, and remove a nearby
redundant call to reserve(). Calling reserve more than once on any
given container instance is always a pessimization.