Add MUMPS to IPOPT when building wheels
Point our wheel build of IPOPT at MUMPS. Because MUMPS is copyleft, we
must consume it as a shared library. On Ubuntu, this means we can just
use the distro-provided package. On macOS, we must build it from source,
as it is not packaged separately in Homebrew. (Homebrew normally builds
it as part of the IPOPT build.)