Fix broken gfortran in macOS wheel build
It appears that gfortran has a somewhat-well-known issue in that it
hard-codes the SDK path that was used to build it, which results in it
falling flat on its face ("ld: library not found for -lm" is especially
popular) if the version of the installed SDK does not match exactly.
While I have never encountered this personally, it appears to affect at
least our Monterey CI machine(s).

Work around this by explicitly setting the SDKROOT to the appropriate
path. (It appears we already do this for Bazel itself, but the macOS
wheel builds need it set when we build dependencies, particularly
MUMPS.)

See also:
https://github.com/xianyi/OpenBLAS/issues/3032#issuecomment-743805371.