[manipulation] Add UR3e model and TRI Homecart
This PR only contains a test that verifies that the model can be
parsed correctly (through model directives, etc). But in a follow-up
PR I will use this as a Drake example of bimanual motion planning with
Graph of Convex Sets.

These models were derived from the Anzu versions; with a few changes:

- Anzu has "left/right" variants of the ur3e models, which set tighter
joint limits for the homecart configuration.  I felt that it was a
little inelegant to put those in the ur3e directory (they should be
homecart specific).  We can add them later if need be.

- I've used the simple wsg gripper models that were already in Drake
instead of the fancier models in anzu.

- I've named the directory `tri_homecart` instead of `homecart`.

- I've manually removed the installed .obj and .png files from the pip
wheels; otherwise it would cause our wheels to exceed the 100MB limit.

- The model files have been optimized by Cody to get the filesize for the homecart-specific model files down to ~ 2MB.