Fix some build errors on Jammy (#17378)
* Fix some build errors on Jammy

Fix some code that's tripping -Werror=range-loop-construct. Fix a
missing include. Fix a -Werror=maybe-uninitialized.

Specifically, new compilers don't like two things: taking a copy of the
loop item where a reference will do, and taking a reference when the
loop item's type is specified such as to force a conversion, since this
is taking a reference to a temporary. (The latter is arguably poor form,
anyway, as it effectively hides a conversion.)