Fix github3_py on Jammy
github3_py requires jwt. It seems something is causing this to be
available on Focal, but on Jammy it needs to be explicitly installed.

Since this is only used by a maintainer tool, the prerequisites are
accordingly maintainer-only. This includes relocating python3-jwcrypto
from test-only to maintainer-only and removing the "test" (which did
little more than verify that the tool doesn't immediately keel over from
an import error).