[pydrake] Defer tkinter import from load-time to ctor-time
The pyplot_visualizer is the only installed module that depends on pyplot
(and thus, depends on tkinter by default on Ubuntu).

We want `import pydrake.all` to provide `PyPlotVisualizer` but we
don't want it to bomb out at import-time for users who never use it.
Instead, we defer the import of pyplot until ctor-time, and give a
friendly error message in case python3-tk is missing.