Reverse the suffix semantics of DrakeLcm::SubscribeAllChannels
This PR proposes to reverse the semantics of DrakeLcm::SubscribeAllChannels
with resepect to channel suffixes.  As documented and tested by my previous
PR, SubscribeAllChannels prior to this PR provides the full, suffix-included
name of the channel.

I assert the prior behaviour was wrong for two reasons:

* First, a caller likely suspects that the channel passed to Subscribe
  and the channel received by SubscribeAllChannels are the same entity.
  For one to be qualified and the other not is a surprise.

* Second, this is very troublesome for clients that receive a DrakeLcm as a
  DrakeLcmInterface, as they cannot determine the suffix with which the
  DrakeLcm was constructed (nor could they even if we added an accessor on
  the subclass).  Thus they cannot know the publisher's intended channel
  name.

This is of course a fairly marginal use case:  SubscribeAllChannels is used
mostly in meta-LCM applications such as logging all messages (or, in my case,
a crude active proxy).  However these meta-LCM applications do care about
the detailed semantics.