Generate .pyi's using pybind11-stubgen
Add pybind11-stubgen as a new external. Add a rule to generate `.pyi`s
for `pydrake` using the same and a small helper script.

For now, this must be manually invoked, and some functions are being
skipped (see #17520). However, this will give us the ability to manually
test the generated `.pyi` files. Adding logic to install them and/or
otherwise bundle them can come later.