PR [geometry] Add test epsilon in preparation for new AutoDiff
The new AutoDiff folds derivative scaling as (b * a) * x' instead of
b * (a * x').  Sometimes this changes the floating point round-off.

The premise of merging this change now is that sensitivity to fold
order means that the test is overly precise, whether or not we ever
merge the new AutoDiff.