PR [primitives] Spell ternary operator compatibly with new AutoDiff
The new AutoDiff will have a breaking change: its derivatives()
mutable accessor will no longer guide the conditional ternary
operator to construct a temporary VectorXd. There is no longer
a "common type" between a constant-expression eigen xpr vector
and an AutoDiff derivatives vector (both a will be eigen xprs,
of different types).

We plan to add sugar to make this easier down the road, but in
any case forcing the constant vector to materialize in test code
is not any great loss, and it's easier for the commit train to
land this now and get it out of the way.