[multibody] Reformulate cassie benchmark case permutations
Rewrite the fixture to be templated on scalar type, remove redundant
member fields, and pre-allocate all function inputs and outputs.

Avoid using zeros for any state or input; math on a zero might have a
different performance profile than non-zero. This is already the case
for Expression, and might be the case for AutoDiff down the road.

For AutoDiff cases, test performance with different sizes of gradient:
- no gradient,
- wrt x only,
- wrt vdot only (for ID cases),
- wrt u only (for FD cases),
- wrt x and vdot (for ID cases),
- wrt x and u (for FD cases).

Similarly for Expression, highlight that the base case has no symbolic
variables and add cases for x, vdot, or u each being symbolic variables
(except that we can't do forward dynamics LDLT on a Variable x).