[pydrake] Fold pydrake.solvers into a single module
The shorter import path and consolidated API docs are more typical of
Python modules than deeply-nested paths, and so are easier for users.

The single shared library also reduces the zipped size of our whl
releases by approximately 7 MiB by reducing the zipped size of the
pydrake.solvers directory by 78%.

Add shim modules so that the old, longer module names continue to work.
Leave the unit tests for pydrake.solvers unchanged (using the old names)
to confirm that the shim modules are operating correctly.

Update tutorials to prefer the new, shorter names. Update tests for
other (non-solvers) modules to prefer the new, shorter names.

Also delete the //bindings/pydrake/solvers:solvers_pybind library target.
This not a breaking change because it's not part of the Stable API.