Suppress valgrind complaint via dlopen (Jammy)
On Jammy, valgrind is reporting an invalid read via dlopen. Since this
comes from the bowels of glibc, it seems unlikely that the error is our
fault, or that we can do anything about it. Thus, suppress it so the
builds will pass (and any "real" errors will better show up).