[math] Reduce implementation boilerplate for autodiff
The clutter was getting in the way of my understanding of the code
while adding support for new autodiff types, and it any case the
simpler versions should be easier to read and maintain overall.

Update project-wide Clang flags to silence its warning about C++20
features imported from the future.

Fix the autodiff initialization and functions to be more aware of
storage order, in terms of documenting what they are doing, preserving
storage order where practical, and static-asserting when misused.

This introduces one breaking change -- ExtractGradient now accounts
for the storage order of its input, instead of always tracing in
ColMajor order. For most users this will be no change because Eigen's
default storage is ColMajor anyway.