Try to avoid mypy timing out (#17584)
Mypy has a timeout in its ModuleInspect that will trip if importing a
module takes an inordinate (>30s) amount of time. This is usually not an
issue, but it seems (possibly due to AWS storage quirks) that we're
hitting this sometimes, more or less at random.

In order to hopefully mitigate this issue, import ourselves (i.e.
pydrake.all) up front, so that everything is hopefully "hot" when mypy
goes to do its importing.