[geometry] Build deformable sdf from distance to surface mesh
Remove the ReferenceDeformableGeometry class and instead stores the
approximate signed distance field directly in deformable geometry. With
this, we unlock arbitrary volume meshes (with potentially non-convex
surface) as deformable geomtries.

While doing that, we ensure that the mesh inside the SDF points to the
deformable mesh held by DeformableGeometry so that we don't have to
create local copies in deformable_mesh_intersection.cc.