[multibody] All bodies and frames must have a non-empty name (#17632)
* [multibody] All bodies and frames must have a non-empty name

Names are not essential for the MbT itself, but they are essential for
all of the related accouterments -- logging, visualization, export, etc.

Rather than make all of that code add special cases for empty names,
it's easier to enforce that names are non-empty.

We could do that by inventing dummy names under the hood (indeed, we
do that here for the deprecation period) but the names will be much
better if we force the user to provide them.