Add mirror for MUMPS (when building wheel)
MUMPS recently yanked (and later restored) the tarball for the older
5.4.0 version (current, as of writing, is 5.5.x), resulting in the CI
wheel builds being broken for two days. In order to be less dependent on
upstream in the future, add our own S3 mirrors as a preferred source for
the tarball.

It's arguable whether we should be doing this for all packages, or
alternatively using e.g. the Ubuntu mirrors. However, most of our other
wheel dependencies come from reputable sites and don't have a known
history of purging old versions. MUMPS expects potential users to fill
out a request form before disclosing the download URI, which is already
dodgy. (We actually obtained the URIs from Homebrew.)