Use Ubuntu mirror for MUMPS (wheel builds)
MUMPS has recently become unreliable as far as providing tarballs. In
order to be less dependent on upstream, switch to using Ubuntu's source
mirror as the preferred source for the tarball.

Note that, because Ubuntu only ships a limited set of versions, this
requires downgrading to 5.4.1, which as far as I can tell has never been
shipped by Homebrew. Thus, in this respect, the macOS wheels will be
more similar to the Ubuntu wheels than the macOS non-wheels.