Deleting the particles example
While the particles example illustrates the absolutely simplest possible
dynamics system (with a non-zero acceleration), that is its *only* value.
Visualizing it shows a sphere instantly racing out of view and the
code manifests an anti-pattern (using AllocateContext() instead of
CreateDefaultContext()).

As there are other examples that introduce bespoke continuous systems
(e.g., pendulum), this examples contribution is not justified.