Auto reconnect to meshcat server
The client will automatically attempt to reconnect to the meshcat url
at a default 1-second cadence. The reconnect period can be tweaked by
changing the url.

   http://localhost:7000/?reconnect_ms=10000

Set the reconnect period to 10,000 ms.

The C++ MeshCat server adds the following to facilitate reconnection:
  - When connecting to a client, it deletes everything in the prefix_
    path.
  - While the server broadcasts the current state to a new connection, any
    value not *explicitly* stored in the current state doesn't get
    broadcast. Properties related to changing render mode have been
    explicitly added to the state. Previously, the initial values were
    implicit.

What this does *not* do: overwrite the lighting.