Updating context-system mismatch error messages
When validating contexts against systems, we've added a bit more analysis
for recognizing whether the system or the context are the root and
provide custom messages reflecting that.

Furthermore, all error messages for Context validation include a URL
in the troubleshooting webpage to provide elaboration.

To that end:
 - Change logic in ValidateContext() for special cases.
 - Change ValidateCreatedForThisSystem() to redirect to ValidateContext()
   for contexts.
 - Write the troubleshooting entry.
 - Tests
   - For dependency reasons, the testing is distributed over two test
     files. system_base_test.cc avoids dependencies on diagram-level
     constructs. So, diagram_test.cc gets those tests.

Of special note: this changes the levels of table of contents on the
webpage. Perusing troubleshooting is greatly enhanced with *three* levels
of TOC. There are only a few pages that have TOC at all, and only *one* had
a header below level 2. So, it's been explicitly omitted and doc/README.txt
has been updated.