[visualization] Auto reconnect to meshcat server (#17685)
The client will automatically attempt to reconnect to the meshcat url
at a default 1-second cadence. The reconnect period can be tweaked by
changing the url.

   http://localhost:7000/?reconnect_ms=10000

Set the reconnect period to 10,000 ms.

Setting the internal to negative will disable the auto reconnect.
Reconnection works by refreshing the page, so all current viewer
state is lost.

When the session is disconnected, a message is displayed
indicating such.