Improve wheel non-platform-agnosticism reliability
Remove various unreliable hacks that attempted to ensure that we don't
build a platform-agnostic wheel and instead always build an extension
module, which seems to be the most reliable way to accomplish this goal.
In order to ensure this works, generate a dummy source file with an
uninteresting symbol so we don't run afoul of compilers griping about
being asked to create a library with no sources.