Modify ApplyVisuailzationConfig -- plant unnecessary for no contact
This changes the semantics to align with the philosophy, "Not having what I
don't need is not an error."

ApplyVisualizationConfig would unilaterally look for a plant in the builder
(whether it needed it or not). The only time it needs the plant if there
is a request to visualize contact. So, now, the search for a plant (and its
concommitant potential error) are put behind a test on whether contact is
requested at all. Not having a plant when unneeded is *not* an error.