[geometry] Maintain DeformableGeometry's signed distance field
As the documentation of
SurfaceVolumeIntersector::SampleVolumeFieldOnSurface states, the volume
field to be sampled from should be based on the volume mesh in its
current configuration, not any arbitrary configuration (e.g. the
reference configuration). This is important because the gradient
information used to interpolate values in the interior of tetrahedra
depends on the configuration of the mesh.

In deformable mesh intersection, we incorrectly used the volume field
based on the reference configuration of the deformable mesh and
consequently were getting garbage penetration distances.