Integrated into build
 - Add entry into BUILD.bazel
 - Update namespaces
 - Change include paths (as appropriate)
 - Update aliases
 - Change name to strip of Drake prefix (an artifact from prepatory work
   in anzu).

Finally, this originally created an OpenGL-based render engine. In anzu,
that was fine (the OpenGL render engine is always available for anzu
applications). For Drake, generally, that is not the case. This swaps
VTK for GL and defers the question of configuring the *type* of render
engine for the future.