[multibody/parsing] Add support for default positions in model directives
Adds the ability to set default joint positions and default free body
poses in the add_model directive. I would have liked to be able to
enable a tag like default_positions which takes a vector, but the
position vector is not defined until after Finalize(), so it
unavailable at parsing time.

Related to #13065, and this slack discussion:
https://drakedevelopers.slack.com/archives/C2WBPQDB7/p1601749507024400