[config] Update examples to use AddDefaultVisualization where convenient
Updates a number of examples visualization implementations to use
AddDefaultVisualization(). There are two groups of such examples:

 - Those that used boilerplate to visualize illustration geometry and
   contacts.
 - Those that only visualized illustration geometry.

The first sort offer a direct mapping to AddDefultVisualization() without
change to functionality. The second were mapped to
AddDefaultVisualization() with the understanding that requesting
visualization of contact is both meaningless and harmless.

In some cases, anachronistic examples were updated to use
AddMultibodySceneGraph() to more closely hew to the recommended spelling.

This excludes numerous binaries in examples (and elsewhere) that defy a
mechanical translation.