[geometry] Improve ComputeSignedDistancePairClosestPoints feedback (#17826)
* Improve ComputeSignedDistancePairClosestPoints feedback

ComputeSignedDistancePairClosestPoints() is the only QueryObject API that
takes GeometryIds. This puts the burden on the user to pass the *right*
GeometryIds (e.g., properly registered with proximity roles). Prior to
this, if they failed, they'd simply get the message:

The geometry given by id 63 does not reference a geometry that can be used
in a signed distance query

This now provides additional insight, indicating *why* it can't be used
and giving insight into why they have the wrong one (e.g., it has an
illustration role instead). This should help guide the user to change
how they select their geometry ids.