[sensors] Add ApplyCameraConfig function (#17827)
This is more or less a verbatim copy of the cleaned up Anzu code. There is
one exception.

ApplyCameraConfig() originally created an OpenGL-based render engine. In
anzu, that was fine (the OpenGL render engine is always available for anzu
applications). For Drake, generally, that is not the case. This swaps
VTK for GL and defers the question of configuring the *type* of render
engine for the future.