[planning] Add initialization event to DiffIK Integrator (#17824)
* [planning] Add initialization event to DiffIK Integrator

The integrator in Diff IK needs to be set to the initial state of the
robot in order to avoid a large delta at time zero. Previously, users
had to call SetPositions() manually; and this has to happen after
Diagram is built (when a context is created, etc). Because this occurs
in a different part of the code from where the Diff IK system is
created and the parameters configured, I often saw people copying part
of the code and forgetting the initialization, leading to strange
errors.

This PR offers a cleaner workflow where an additional port is
connected at the time that DiffIK is constructed and added to the
builder; then an Initialization event takes care of setting the
initial integrator state.