[parsing] Restructure parsers for format-agnostic recursion
Make parser selection a general mechanism, so that (in future) model
directives will not have to depend on Parser APIs directly. This is a
preparation for making model directives files usable with the regular
Parser APIs.

As a consesquence, this mechanism in principle allows files of any
format to consume files of any other format, using Drake's parsers. This
PR does not attempt to expand the currently supported combinations.

Also, public documentation is reorganized, in anticipation of support
for more formats.